requirements.txt
amqp==2.1.1
appdirs==1.4.0
args==0.1.0
arrow==0.10.0
astor==0.5
autoflake==0.6.6
billiard==3.5.0.2
boto==2.43.0
cached-property==1.3.0
celery==4.0.0
cffi==1.9.1
chardet==2.3.0
chromedriver==2.24.1
clint==0.5.1
colorama==0.3.7
cryptography==1.7.2
cupshelpers==1.0
Django==1.10.6
django-appconf==1.0.2
django-compressor==2.1
django-daterange-filter==1.3.0
django-debug-toolbar==1.6
django-extensions==1.7.4
django-formtools==1.0
django-ordered-model==1.3.0
django-otp==0.3.7
django-phonenumber-field==0.7.2
django-photologue==3.6
django-rainbowtests==0.6.0
django-reversion==2.0.7
django-sortedm2m==1.3.2
django-two-factor-auth==1.4.0
docker==2.1.0
docker-compose==1.11.2
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
ecdsa==0.13
ExifRead==2.1.2
Fabric3==1.13.1.post1
fuzzywuzzy==0.15.0
google-api-python-client==1.5.5
gufw==17.4.1
httplib2==0.9.2
hy==0.12.1
idna==2.2
jedi==0.10.0
json-rpc==1.10.3
jsonschema==2.6.0
kombu==4.0.0
lightdm-gtk-greeter-settings==1.2.1
louis==3.0.0
mypy-lang==0.4.6
npyscreen==4.10.5
oauth2client==4.0.0
packaging==16.8
pacman-mirrors==3.2.2
paramiko==2.1.1
phonenumbers==7.7.0
Pillow==3.4.1
psycopg2==2.6.2
py3status==3.4
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.5.0
Pygments==2.2.0
pygobject==3.22.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-Levenshtein==0.12.0
pytz==2016.6.1
PyYAML==3.12
qrcode==4.0.4
ranger==1.8.1
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
rjsmin==1.0.12
rply==0.7.4
rsa==3.4.2
selenium==3.0.2
service-factory==0.1.5
six==1.10.0
speedtest-cli==1.0.2
sqlparse==0.2.1
team==1.0
texttable==0.8.7
uritemplate==3.0.0
urllib3==1.20
vine==1.1.3
virtualenv==15.1.0
virtualfish==1.0.5
websocket-client==0.40.0
Werkzeug==0.12.1
